Removing the localization module.
For future reference, here are the changes I made:

Replaced:

* localization.exceptions.CcccRecordError with BufferError
* localization.exceptions.ComponentLinkingFailure with KeyError
* localization.exceptions.ConsistencyError with ValueError
* localization.exceptions.InputInspectionDiscontinued with KeyboardInterrupt
* localization.exceptions.InputInspectionMalformed with RuntimeError
* localization.exceptions.InvalidSelectionError with KeyError
* localization.exceptions.IsotxsError, GamisoError, CompxsError, PmatrxError with OSError
* localization.exceptions.NegativeComponentArea, NegativeComponentVolume with ArithmeticError
* localization.exceptions.OverConfiguredError with RuntimeError
* localization.exceptions.RunLogPromptCancel with KeyboardInterrupt
* localization.exceptions.RunLogPromptUnresolvable with RuntimeError
* localization.exceptions.StateError with RuntimeError
* localization.exceptions.SymmetryError with LookupError
* localization.exceptions.SynchronizationError with ValueError
* localization.exceptions.XSLibraryError with AttributeError

Moved:

* latticePhysics_SkippingXsGen_BuChangedLessThanTolerance to latticePhysicsInterface.py
* localization.exceptions to armi.utils.customExceptions
* LatticePhysicsWriter_Nuclide_name_FoundMultipleTimes to latticePhysicsWriter.py
* nuclide_NuclideLabelDoesNotMatchNuclideLabel to xsNuclides.py
* textProcessors to armi/utils/textProcessors.py
* general_submittedJobs to armi.utils.customExceptions
* gereral_waitingForJobs to armi.utils.customExceptions

Removed:

* general_single_TooManyOrTooFewMatchesFor
* localization.exceptions.RangeError
* nuclides_TooManyOrTooFew_number_MatchesForNuclide_name
