+++ armi/physics/neutronics/diffIsotxs.py
@@ -71,7 +71,7 @@ class CompareIsotxsLibraries(EntryPoint):
 #     flux = []
 #     try:
 #         flux_pattern = re.compile(r'^\s*\d+ +(?P<energy>{0}) +{0} +(?P<flux>{0})'
-#                                   .format(strings.SCIENTIFIC_PATTERN))
+#                                   .format(SCIENTIFIC_PATTERN))
 #         energy_iter = iter(energy_groups)
 #         with open(args.flux_file) as flux_bg:
 #             for line in flux_bg:
