+++ armi/tests/test_apps.py
@@ -21,7 +21,6 @@ import unittest
 
 import armi
 from armi import plugins
-from armi.localization import exceptions
 from armi.__main__ import main
 
 
@@ -122,7 +121,7 @@ class TestArmi(unittest.TestCase):
         self.assertIn(cli.EntryPointsPlugin, pm.get_plugins())
 
     def test_overConfigured(self):
-        with self.assertRaises(exceptions.OverConfiguredError):
+        with self.assertRaises(RuntimeError):
             armi.configure()
 
     def test_main(self):
